Fix Win64 build error.
R=dalecurtis@chromium.org
TBR=dalecurtis@chromium.org

Review URL: https://codereview.chromium.org/70873002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@234707 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e1085ae8319b98dd28c6f27c4232b25909f620b5

+++ media/cast/test/transport/transport.cc
@@ -103,7 +103,8 @@ class LocalPacketSender : public PacketSender {
     scoped_refptr<net::WrappedIOBuffer> buffer(
         new net::WrappedIOBuffer(reinterpret_cast<const char*>(data)));
     int rv = udp_socket_->SendTo(
-        buffer.get(), packet.size(), send_address_, callback.callback());
+        buffer.get(), static_cast<int>(packet.size()), send_address_,
+        callback.callback());
     return (rv == packet.size());
   }
